webapp/requirements.txt
Django==4.0.2
djangorestframework==3.13.1
celery==5.2.3